requirements.txt
kafka-python==2.0.1
boto3==1.12.26